Add option/setting to prevent declaration of queue (#37)
In certain setups, queue will already be declared when a consumer gets started. Starting the consumer will fail if the declared queue and the consumers settings mismatch. In order to prevent the need for managing the queue settings in multiple places, one might want to prevent the consumer from declaring the queue. This change adds a new cli option `-no-declare` and a new setting `nodeclare` to the `[queuesettings]` section of the config file which allows to do exactly this.

resolves #36.